--- src/main/java/com/github/lostizalith/velka/account/AccountService.java
@@ -1,5 +1,0 @@
-package com.github.lostizalith.velka.account;
-
-public interface AccountService {
-
-}
