+++ src/main/java/com/github/lostizalith/velka/account/entity/AccountCurrency.java
@@ -1,4 +1,4 @@
-package com.github.lostizalith.velka.account;
+package com.github.lostizalith.velka.account.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
