+++ src/main/java/com/github/lostizalith/velka/account/service/AccountService.java
@@ -1,0 +1,5 @@
+package com.github.lostizalith.velka.account.service;
+
+public interface AccountService {
+
+}
