--- src/main/java/com/github/lostizalith/velka/category/CategoryService.java
@@ -1,5 +1,0 @@
-package com.github.lostizalith.velka.category;
-
-public interface CategoryService {
-
-}
