+++ src/main/java/com/github/lostizalith/velka/category/controller/CategoryController.java
@@ -1,5 +1,6 @@
-package com.github.lostizalith.velka.record;
+package com.github.lostizalith.velka.category.controller;
 
+import com.github.lostizalith.velka.category.service.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
@@ -8,8 +9,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
-@RequestMapping(path = "/api/v1/records", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-public class RecordController {
+@RequestMapping(path = "/api/v1/categories", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+public class CategoryController {
 
-    private final RecordService recordService;
+    private final CategoryService categoryService;
 }
