+++ src/main/java/com/github/lostizalith/velka/category/service/CategoryService.java
@@ -1,0 +1,5 @@
+package com.github.lostizalith.velka.category.service;
+
+public interface CategoryService {
+
+}
