--- src/main/java/com/github/lostizalith/velka/record/RecordService.java
@@ -1,5 +1,0 @@
-package com.github.lostizalith.velka.record;
-
-public interface RecordService {
-
-}
