+++ src/main/java/com/github/lostizalith/velka/record/service/RecordService.java
@@ -1,0 +1,5 @@
+package com.github.lostizalith.velka.record.service;
+
+public interface RecordService {
+
+}
